# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GoBA Slim API
  version: 0.1.0
...
paths:
...
  /api/v1/system-configs:
    get:
      tags: [SystemConfig]
      operationId: listSystemConfigs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System configs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfigList'
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
    post:
      tags: [SystemConfig]
      operationId: createSystemConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSystemConfigRequest'
      responses:
        '201':
          description: System config created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SystemConfig'}
        '400': {$ref: '#/components/responses/ErrorResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
...
  /api/v1/files:
    post:
      tags: [Files]
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileObject'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '415':
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/files/{ownerId}/{fileName}:
    parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: fileName
        in: path
        required: true
        schema:
          type: string
          pattern: '^[0-9a-f-]{36}\.(jpg|png|gif|webp|mp4|webm)$'
    delete:
      tags: [Files]
      operationId: deleteFile
      security:
        - bearerAuth: []
      responses:
        '204':
          description: File deleted
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /files/{ownerId}/{fileName}:
    parameters:
      - name: ownerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: fileName
        in: path
        required: true
        schema:
          type: string
          pattern: '^[0-9a-f-]{36}\.(jpg|png|gif|webp|mp4|webm)$'
    get:
      tags: [Files]
      operationId: getFile
      responses:
        '200':
          description: Public file
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            image/jpeg: {schema: {type: string, format: binary}}
            image/png: {schema: {type: string, format: binary}}
            image/gif: {schema: {type: string, format: binary}}
            image/webp: {schema: {type: string, format: binary}}
            video/mp4: {schema: {type: string, format: binary}}
            video/webm: {schema: {type: string, format: binary}}
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    FileObject:
      type: object
      required: [key, url, content_type, size]
      properties:
        key:
          type: string
        url:
          type: string
          format: uri-reference
        content_type:
          type: string
        size:
          type: integer
          format: int64
    SystemConfigValueType:
      type: string
      enum: [string, integer, boolean, duration, string_list]
    CreateSystemConfigRequest:
      type: object
      required: [key, value, value_type, is_public]
      properties:
        key: {type: string}
        value: {}
        value_type: {$ref: '#/components/schemas/SystemConfigValueType'}
        is_public: {type: boolean}
        description: {type: string, maxLength: 255}
...
    SystemConfig:
      type: object
      required: [key, value, value_type, is_public, description, created_at, updated_at]
      properties:
        key: {type: string}
        value: {}
        value_type: {$ref: '#/components/schemas/SystemConfigValueType'}
        is_public: {type: boolean}
        description: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    SystemConfigList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/SystemConfig'}
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...