# db/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type SystemConfig struct {
	Key         string
	Value       []byte
	ValueType   string
	IsPublic    bool
	Description string
	CreatedAt   pgtype.Timestamptz
	UpdatedAt   pgtype.Timestamptz
}

type User struct {
	ID                    uuid.UUID
	Username              string
	PasswordHash          string
	DisplayName           string
	Email                 pgtype.Text
	AvatarUrl             pgtype.Text
	Status                string
	IsSuperuser           bool
	AllowMultipleSessions bool
	SessionVersion        int64
	PasswordChangedAt     pgtype.Timestamptz
	LastLoginAt           pgtype.Timestamptz
	CreatedAt             pgtype.Timestamptz
	UpdatedAt             pgtype.Timestamptz
	ArchivedAt            pgtype.Timestamptz
}